Cập nhật tính năng thêm mới card trong 1 list
