+++ src/app/board/board.service.ts
@@ -3,20 +3,26 @@ import {BoardModel} from '../shared/board-create/board-model';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {LocalStorageService} from 'ngx-webstorage';
 import {Observable} from 'rxjs';
+import {Route, Router} from '@angular/router';
+import {GBoard} from '../gboard';
 
 @Injectable({
   providedIn: 'root'
 })
 export class BoardService {
 
-  constructor(private http: HttpClient,
+  constructor(private httpClient: HttpClient,
+              private router: Router,
               private localStorage: LocalStorageService) { }
 
   // tslint:disable-next-line:typedef
   createBoard(newBoard: BoardModel): Observable<Object> {
     // const header = new HttpHeaders().set('Authorization', 'Bearer ' + this.localStorage.retrieve('authenticationToken'));
-    return this.http.post('http://localhost:8080/api/boards/', newBoard);
+    return this.httpClient.post('http://localhost:8080/api/boards/', newBoard);
   }
 
-  listCard
+  getBoardList(): Observable<GBoard[]>{
+    const id = this.localStorage.retrieve('userId');
+    return this.httpClient.get<GBoard[]>('http://localhost:8080/api/users/' + id);
+  }
 }
