+++ src/app/card/card.service.ts
@@ -11,7 +11,6 @@ export class CardService {
   constructor(private httpClient: HttpClient) { }
 
   creatCard(newCard: GCard): Observable<Object>{
-    console.log(newCard);
     return this.httpClient.post('http://localhost:8080/api/cards/', newCard);
   }
